feat(#79):corrigindo estrutura e caminhos Corrigindo estrutura e caminhos do código
