+++ src/app/home/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useSession } from 'next-auth/react';
-import JourneyPage from '@/app/components/home/journeyPage';
+import HomePrincipalPage from '@/components/home/homePage';
 
 
 export default function HomePage() {
@@ -10,7 +10,7 @@ export default function HomePage() {
   return (
     <div className='bg-[#F1F1F1] min-h-screen'>
       <div className='bg-[#F1F1F1] max-w-7xl mx-auto pt-14 px-5'>
-          <JourneyPage />
+          <HomePrincipalPage />
       </div>
     </div>
   );
