+++ src/components/home/homeJourneyCard.tsx
