+++ src/components/sidebar.component.tsx
@@ -1,47 +1,20 @@
 'use client';
 
 import Link from 'next/link';
-import { useSession } from 'next-auth/react';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import { Drawer, IconButton, Box } from '@mui/material';
 import HomeIcon from '@mui/icons-material/Home';
+import DashboardIcon from '@mui/icons-material/Dashboard';
+import EditNoteIcon from '@mui/icons-material/EditNote';
+import { Drawer, IconButton, Box } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
-import FollowTheSignsIcon from '@mui/icons-material/FollowTheSigns';
+import { useSession } from 'next-auth/react';
 
 interface SideBarProps {
   handleDrawerOpen: () => void;
   open: boolean;
 }
 
-interface SidebarItem {
-  label: string;
-  href: string;
-  icon: React.ReactNode;
-  roles?: string[];
-}
-
-const sidebarItems: SidebarItem[] = [
-  {
-    label: 'Home',
-    href: '/home',
-    icon: <HomeIcon className="h-5 w-5 mr-2" />,
-  },
-  {
-    label: 'Painel de Administrador',
-    href: '/admin',
-    icon: <DashboardIcon className="h-5 w-5 mr-2" />,
-    roles: ['admin'],
-  },
-  {
-    label: 'Minhas Jornadas',
-    href: '/journeys',
-    icon: <FollowTheSignsIcon className="h-5 w-5 mr-2" />,
-    roles: ['teacher', 'admin'],
-  },
-];
-
 const Sidebar: React.FC<SideBarProps> = ({ handleDrawerOpen, open }) => {
-  const { data: session } = useSession();
+  const session = useSession();
 
   return (
     <Drawer anchor="left" open={open} onClose={handleDrawerOpen}>
@@ -50,26 +23,35 @@ const Sidebar: React.FC<SideBarProps> = ({ handleDrawerOpen, open }) => {
           <CloseIcon />
         </IconButton>
         <ul>
-          {sidebarItems
-            .filter((item) =>
-              session && item.roles
-                ? item.roles.includes(session.user.role)
-                : true,
-            )
-            .map((item) => (
-              <li
-                key={item.href}
-                className="mb-2 flex items-center p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200"
+          <li className="mb-2 flex items-center p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200">
+            <HomeIcon className="h-5 w-5 mr-2" />
+            <Link
+              href="/home"
+              className="block p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200"
+            >
+              Home
+            </Link>
+          </li>
+          {session.data?.user.role === 'admin' && (
+            <li className="mb-2 flex items-center p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200">
+              <DashboardIcon className="h-5 w-5 mr-2" />
+              <Link
+                href="/admin"
+                className="block p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200"
               >
-                {item.icon}
-                <Link
-                  href={item.href}
-                  className="block p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200"
-                >
-                  {item.label}
-                </Link>
-              </li>
-            ))}
+                Painel de Administrador
+              </Link>
+            </li>
+          )}
+          <li className="mb-2 flex items-center p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200">
+            <EditNoteIcon className="h-5 w-5 mr-2" />
+            <Link
+              href="/studio"
+              className="block p-2 hover:bg-blue-100 hover:text-purple-600 transition duration-200"
+            >
+              Estúdio de Criação
+            </Link>
+          </li>
         </ul>
       </Box>
     </Drawer>
